Cast `request_timeout` to `int`

+++ api-server/config/openai.php
@@ -3,5 +3,5 @@
 return [
     'api_key' => env('OPENAI_API_KEY'),
     'organization' => env('OPENAI_ORGANIZATION'),
-    'request_timeout' => env('OPENAI_REQUEST_TIMEOUT', 30),
+    'request_timeout' => (int) env('OPENAI_REQUEST_TIMEOUT', 30),
 ];
